feat: Make qr code scanning modular
